requirements.txt
streamlit
opencv-python
opencv-contrib-python
numpy
pandas
Pillow
